+++ src/lightly_train/_task_models/ltdetr_object_detection/config.py
