+++ src/lightly_train/_task_models/ltdetr_object_detection/dinov3_convnext_wrapper.py
