+++ src/lightly_train/_task_models/ltdetr_object_detection/dinov3_vit_wrapper.py
