+++ src/lightly_train/_task_models/ltdetr_object_detection/train_model.py
@@ -35,16 +35,16 @@ from lightly_train._metrics.detection.task_metric import (
     ObjectDetectionTaskMetricArgs,
 )
 from lightly_train._optim import optimizer_helpers
-from lightly_train._task_models.dinov3_ltdetr_object_detection.dinov3_vit_wrapper import (
+from lightly_train._task_models.ltdetr_object_detection.dinov3_vit_wrapper import (
     DINOSTAs,
 )
-from lightly_train._task_models.dinov3_ltdetr_object_detection.ecvit_vit_wrapper import (
+from lightly_train._task_models.ltdetr_object_detection.ecvit_vit_wrapper import (
     ECViTBackboneWrapper,
 )
-from lightly_train._task_models.dinov3_ltdetr_object_detection.task_model import (
+from lightly_train._task_models.ltdetr_object_detection.task_model import (
     LTDETRObjectDetection,
 )
-from lightly_train._task_models.dinov3_ltdetr_object_detection.transforms import (
+from lightly_train._task_models.ltdetr_object_detection.transforms import (
     LTDETRObjectDetectionTrainTransform,
     LTDETRObjectDetectionTrainTransformArgs,
     LTDETRObjectDetectionValTransform,
